Added HTML code with additional formatting.
Type: Code Addition

Description: This commit adds a snippet of HTML code that showcases different forms of additional formatting on text elements. The HTML code includes the use of the <mark> tag to highlight text, the <small> tag to display text in a smaller size, the <del> tag to display strikethrough text, the <ins> tag to display underlined text, and the <sup> and <sub> tags to display superscript and subscript text, respectively. The HTML code has been tested and correctly displays the additional formatting. The elements have been successfully added to the repository.