Added quotations and HTML code.
Type: Code Addition.

Description: This commit adds a snippet of HTML code that showcases the usage of quotations and code elements. The HTML code includes the use of the <q> tag to highlight a quotation, <code> and <pre> tags to enclose a block of code with preformatted styling, <blockquote> tag to display a blockquote in a highlighted manner, <abbr> tag to show an abbreviation with a title, and <bdo> tag to display text in reverse order. The HTML code has been tested and correctly displays the quotations and code. The elements have been successfully added to the repository.