Added Ordered Lists (OL) and Unordered Lists (UL).
Type: Code Addition

Description: This commit adds HTML code snippets that demonstrate the usage of ordered lists (OL) and unordered lists (UL). The HTML code includes the use of the <ol> tag to create an ordered list with multiple items (using the <li> tag for each item) and the <ul> tag to create an unordered list with multiple items, also using the <li> tag for each item. The HTML code has been tested and correctly displays the ordered and unordered lists. The elements have been successfully added to the repository.