Added Links and Anchors in HTML5.
Type: Code Addition

Description: This commit adds an HTML code snippet that demonstrates the usage of links and anchors in HTML5. The HTML code includes an anchor element (<a>) with an "href" attribute pointing to the URL "https://github.com/NanamiWatanabe". The "target" attribute is set to "_blank", which opens the link in a new browser tab or window. Additionally, the "rel" attribute is set to "external", indicating that the link leads to an external website. The text "Lorem ipsum" is displayed as the link text. The HTML code has been tested, and the link functions correctly. The code snippet has been successfully added to the repository.