Added Internal Links in HTML5.
Type: Code Addition

Description: This commit adds two HTML files that demonstrate the usage of internal links in HTML5. The "index.html" file contains a paragraph with two anchor elements (<a>). The first anchor has an "href" attribute pointing to the "next.html" file, and the "rel" attribute is set to "prev", indicating it is a link to the previous page. The second anchor has an "href" attribute also pointing to the "next.html" file, and the "rel" attribute is set to "nofollow". The text "Lorem ipsum dolor" is displayed as the link text. The "next.html" file contains a paragraph with an anchor that has an "href" attribute pointing to the "index.html" file, and the "rel" attribute is set to "next", indicating it is a link to the next page. The text "Lorem ipsum dolor" is displayed as the link text. The HTML files have been tested, and the links function correctly. The files have been successfully added to the repository.