Added Download Links in HTML5.
Type: Code Addition

Description: This commit adds an HTML file that demonstrates the usage of download links in HTML5. The paragraph contains two anchor elements (<a>). The first anchor has an "href" attribute pointing to the "Lorem Ipsum.pdf" file in the "download/" folder, and the "download" attribute is set as "Lorem Ipsum.pdf", indicating the filename to be downloaded. The "type" attribute is set as "application/pdf", specifying the MIME type of the file. The text "Lorem ipsum dolor sit" is displayed as the link text. The second anchor has an "href" attribute pointing to the "Lorem Ipsum.zip" file in the "download/" folder, and the "download" attribute is set as "Lorem Ipsum.zip". The "type" attribute is set as "application/zip". The text "veritatis" is displayed as the link text. The PDF and ZIP files are available for download. The HTML file has been tested, and the download links function correctly. The file has been successfully added to the repository.