Added Social Media Accounts.
Type: Code Addition

Description: This commit adds an HTML file that displays information about an anime character named Haruka Mizuki and her fictional social media accounts. The text describes Haruka as a brilliant student and powerful sorceress who controls the elements of fire, water, and wind. Additionally, she is skilled in swordsmanship. The social media accounts are displayed as an unordered list (UL) with links to each account. Each link has an "href" attribute pointing to the respective social media address, and the "target" and "rel" attributes are set to "_blank" and "external" respectively, to open the links in a new window and indicate that they are external links. The social media icons are displayed as images within the links. The HTML file has been tested, and the links function correctly. The file has been successfully added to the repository.