Added self-adapting images.
Type: Code Addition

Description: This commit adds an HTML file that includes a title "Self-Adapting Images" and a sample paragraph of text. The file also includes a "picture" element that contains three different image sources, each with a corresponding media query to adjust the image based on the screen width. For screen widths up to 750px, the displayed image is "img/300x300.png". For screen widths up to 1050px, the displayed image is "img/700x700.png". For screen widths above 1050px, the displayed image is "img/1000x1000.png". The commit successfully adds the HTML file to the repository.