Added video formats to the website.
Type: Code Addition

Description: This commit adds an HTML file that includes a title "Added video formats to the website", a paragraph of text, and a "video" element that loads a video file "video/LoremIpsum.mp4". The "controls" attribute adds playback controls for the user to interact with the video. The "width" attribute sets the width of the video to 300 pixels. The "poster" attribute specifies a poster image that will be displayed before playing the video. The commit successfully adds the HTML file to the repository.